Move methods to ai.cpp
