+++ variables.hpp
@@ -26,6 +26,14 @@ ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
 
+// ai.cpp
+ELONA_EXTERN(int cellchara);
+ELONA_EXTERN(int cellfeat);
+ELONA_EXTERN(elona_vector1<int> isetthrowpotionminor);
+ELONA_EXTERN(elona_vector1<int> isetthrowpotionmajor);
+ELONA_EXTERN(elona_vector1<int> isetthrowpotiongreater);
+ELONA_EXTERN(int npccostmp);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -555,7 +563,6 @@ bool actionsp(int, int);
 int addbuff(int = 0, int = 0, int = 0, int = 0);
 int advfavoriteskill(int = 0);
 int advfavoritestat(int = 0);
-int ai_check();
 int blendcheckext(int = 0, int = 0);
 int blendcheckmat(int = 0);
 int blendmatnum(int = 0, int = 0);
@@ -1351,9 +1358,6 @@ void complete_quest();
 void do_play_scene();
 void label_2685();
 turn_result_t npc_turn();
-turn_result_t proc_misc_npc_map_events();
-turn_result_t proc_ally_town_training(bool = false);
-turn_result_t proc_ai_basic();
 void show_quick_menu();
 void show_ex_help();
 void label_2702();
